+++ examples/ball/sim.py
@@ -39,9 +39,7 @@ def apply_wind(
     def apply_wind_inner(f, v):
         v_diff = w[0] - v.linear()
         v_diff_dir = v_diff / la.norm(v_diff)
-        return el.SpatialForce.from_linear(
-            f.force() + 0.2 * 0.5 * v_diff**2 * v_diff_dir
-        )
+        return el.SpatialForce(linear=f.force() + 0.2 * 0.5 * v_diff**2 * v_diff_dir)
 
     return q.map(
         el.Force,
@@ -51,17 +49,15 @@ def apply_wind(
 
 @el.map
 def gravity(f: el.Force, inertia: el.Inertia) -> el.Force:
-    return f + el.SpatialForce.from_linear(
-        jnp.array([0.0, 0.0, -9.81]) * inertia.mass()
-    )
+    return f + el.SpatialForce(linear=jnp.array([0.0, 0.0, -9.81]) * inertia.mass())
 
 
 @el.map
 def bounce(p: el.WorldPos, v: el.WorldVel) -> el.WorldVel:
     return jax.lax.cond(
         jax.lax.max(p.linear()[2], v.linear()[2]) < 0.0,
-        lambda _: el.SpatialMotion.from_linear(
-            v.linear() * jnp.array([1.0, 1.0, -1.0]) * 0.85
+        lambda _: el.SpatialMotion(
+            linear=v.linear() * jnp.array([1.0, 1.0, -1.0]) * 0.85
         ),
         lambda _: v,
         operand=None,
@@ -73,9 +69,7 @@ def world(seed: int = 0) -> el.World:
     world.spawn(Globals(seed=jnp.int64(seed)), name="Globals")
     world.spawn(
         [
-            el.Body(
-                world_pos=el.SpatialTransform.from_linear(jnp.array([0.0, 0.0, 6.0]))
-            ),
+            el.Body(world_pos=el.SpatialTransform(linear=jnp.array([0.0, 0.0, 6.0]))),
             world.shape(el.Mesh.sphere(0.4), el.Material.color(12.7, 9.2, 0.5)),
         ],
         name="Ball",